Move tests that dont belong in test_offsets
